pyproject.toml
[project]
name = "detector-neural-network"
version = "1.0.0"
description = "Find spectral lines in recorded data with the help of a neural network"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "joblib",
    "numpy>=1.26.0,<2.1.0",
    "pandas",
    "pyqtgraph",
    "scipy",
    "scikit-learn",
    "PySide6-Essentials",
]

[project.scripts]
detect = "main:main"

[tool.setuptools]
package-dir = { "" = "." }
package-data = {"*" = ["*.py", "*.qss", "*.qrc", "*.svg", "*.png"], "color_theme" = ["*.qss", "*.py"], "resource" = ["*.svg", "*.png"]}